separate out argparsing
