+++ deepcell_applications/app_runners_test.py
@@ -25,7 +25,6 @@
 # ==============================================================================
 """Tests for deepcell_applications.app_runners"""
 import os
-import tempfile
 
 import skimage.io as io
 import numpy as np
@@ -35,54 +34,43 @@ import pytest
 import deepcell_applications as dca
 
 
-def test_run_app_mesmer(mocker):
-    with tempfile.TemporaryDirectory() as temp_dir:
-        # run with default parameters
+def test_run_app_mesmer(tmpdir):
+    temp_dir = str(tmpdir)
 
-        # create required input files and directories
-        img = np.zeros((10, 10))
-        output_dir = os.path.join(temp_dir, 'output_dir')
-        img_path = os.path.join(temp_dir, 'img.tiff')
+    # run with default parameters
 
-        io.imsave(img_path, img)
-        os.makedirs(output_dir)
+    # create required input files and directories
+    img = np.zeros((10, 10))
+    output_dir = os.path.join(temp_dir, 'output_dir')
+    img_path = os.path.join(temp_dir, 'img.tiff')
 
-        # patch user supplied arguments
-        class mocked_get_args_mesmer():
-            def parse_args(self):
-                required_inputs = ['mesmer',
-                                   '--output-directory', output_dir,
-                                   '--nuclear-image', img_path,
-                                   '--squeeze'
-                                   ]
-                formatted_args = dca.argparse.get_arg_parser().parse_args(required_inputs)
-                return formatted_args
+    io.imsave(img_path, img)
+    os.makedirs(output_dir)
 
-        mocker.patch('deepcell_applications.app_runners.get_arg_parser',
-                     mocked_get_args_mesmer)
+    required_inputs = ['mesmer',
+                       '--output-directory', output_dir,
+                       '--nuclear-image', img_path,
+                       '--squeeze'
+                       ]
+    args = dca.argparse.get_arg_parser().parse_args(required_inputs)
 
-        dca.app_runners.run_application()
+    dca.app_runners.run_application(dict(args._get_kwargs()))
 
-        # error checking
+    # error checking
 
-        # create required input files and directories
-        out_path = os.path.join(temp_dir, 'out_mask.tiff')
-        io.imsave(out_path, img)
+    # create required input files and directories
+    out_path = os.path.join(temp_dir, 'out_mask.tiff')
+    io.imsave(out_path, img)
 
-        # patch user supplied arguments
-        class mocked_get_args_mesmer_existing_file():
-            def parse_args(self):
-                required_inputs = ['mesmer',
-                                   '--output-directory', output_dir,
-                                   '--nuclear-image', img_path,
-                                   '--output-name', out_path,
-                                   '--squeeze'
-                                   ]
-                formatted_args = dca.argparse.get_arg_parser().parse_args(required_inputs)
-                return formatted_args
+    # patch user supplied arguments
 
-        mocker.patch('deepcell_applications.app_runners.get_arg_parser',
-                     mocked_get_args_mesmer_existing_file)
+    required_inputs = ['mesmer',
+                       '--output-directory', output_dir,
+                       '--nuclear-image', img_path,
+                       '--output-name', out_path,
+                       '--squeeze'
+                       ]
+    args_io_error = dca.argparse.get_arg_parser().parse_args(required_inputs)
 
-        with pytest.raises(IOError):
-            dca.app_runners.run_application()
+    with pytest.raises(IOError):
+        dca.app_runners.run_application(dict(args_io_error._get_kwargs()))
